requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-groq
python-dotenv
pandas
numpy
matplotlib
seaborn
SQLAlchemy
tabulate
